move folders into constants
