+++ app/constants/folders.ts
@@ -1,0 +1,6 @@
+import os from "os";
+
+const homedir = os.homedir();
+export const progFolder = `${homedir}/.vex-desktop`;
+export const dbFolder = `${progFolder}/databases`;
+export const keyFolder = `${progFolder}/keys`;
